グループ単位での試合結果をCSVエクスポート/インポートする機能を追加 (#242)
* CSVエクスポート機能を実装し，グループ詳細画面にCSVエクスポートボタンを追加．

* グループCSVファイルのフォーマットを変更．グループ情報をファイル先頭に配置．

* グループ一覧画面にインポートボタンを追加し，対応すエンドポイントを定義．内部の実装は未完了．

* アーカイブされたグループに属する試合レコードの状態設定の誤りを修正 (#240)

* アーカイブされたグループに対しては，詳細画面で試合レコードを操作できないようにインタフェースを変更． (#241)

* CSVのインポート機能を実装

* CSVインポート時のチームバージョンの誤りを修正

* チームアーカイブが存在しないチーム情報を削除するときに一斉削除されるバグを修正

+++ server/team/views.py
@@ -313,11 +313,12 @@ def delete_teams():
             return redirect(url_for("team.show_archived"))
 
         # Delete the archive file
-        abs_path = os.path.join(current_app.static_folder, os.path.dirname(team.archive_path))
-        if os.path.exists(abs_path):
-            # print(f"Delete {abs_path}")
-            shutil.rmtree(abs_path)
-            current_app.logger.info(f"Deleted {abs_path}")
+        if team.archive_path is not None and team.archive_path != "":
+            abs_path = os.path.join(current_app.static_folder, os.path.dirname(team.archive_path))
+            if os.path.exists(abs_path):
+                # print(f"Delete {abs_path}")
+                shutil.rmtree(abs_path)
+                current_app.logger.info(f"Deleted {abs_path}")
 
         db.session.delete(team)
         db.session.commit()
@@ -408,6 +409,10 @@ def download(name, version):
     team = Team.query.filter_by(name=name, version=version).first()
     if team:
         abs_path = os.path.join(current_app.static_folder, team.archive_path)
+        if not os.path.exists(abs_path) or not os.path.isfile(abs_path):
+            flash(f"Archive file [{team.archive_path}] not found.", "error")
+            current_app.logger.error(f"download: Archive file [{abs_path}] not found.")
+            return redirect(url_for("team.index"))
         try:
             return send_file(abs_path, as_attachment=True)
         except FileNotFoundError:
